budgeting app
This is a web-based personal budgeting application that allows users to track their income and expenses. Users can create accounts and log in to access their budget overview and add, edit, and delete their income and expenses. The app displays the total income, total expenses, and remaining budget for the user. The back end of the app is built with Java and the front end is built with React